refactor(api): move certification-center-invitation admin route
to src
