+++ api/tests/team/acceptance/application/certification-center-invitation/certification-center-invitation.admin.route.test.js
@@ -14,6 +14,66 @@ describe('Acceptance | Team | Application | Route | Admin | Certification Center
     server = await createServer();
   });
 
+  describe('GET /api/admin/certification-centers/{certificationCenterId}/invitations', function () {
+    it('should return 200 HTTP status code and the list of invitations', async function () {
+      // given
+      const adminUser = await insertUserWithRoleSuperAdmin();
+
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      const now = new Date();
+      const certificationCenterInvitation1 = databaseBuilder.factory.buildCertificationCenterInvitation({
+        certificationCenterId,
+        status: CertificationCenterInvitation.StatusType.PENDING,
+        email: 'alex.terieur@example.net',
+        role: 'MEMBER',
+        updatedAt: now,
+      });
+      const certificationCenterInvitation2 = databaseBuilder.factory.buildCertificationCenterInvitation({
+        certificationCenterId,
+        status: CertificationCenterInvitation.StatusType.PENDING,
+        email: 'sarah.pelle@example.net',
+        role: 'ADMIN',
+        updatedAt: now,
+      });
+      databaseBuilder.factory.buildCertificationCenterInvitation({
+        certificationCenterId,
+        status: CertificationCenterInvitation.StatusType.ACCEPTED,
+      });
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'GET',
+        url: `/api/admin/certification-centers/${certificationCenterId}/invitations`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(adminUser.id) },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.result.data.length).to.equal(2);
+      expect(response.result.data).to.deep.have.members([
+        {
+          type: 'certification-center-invitations',
+          id: certificationCenterInvitation1.id.toString(),
+          attributes: {
+            email: 'alex.terieur@example.net',
+            role: 'MEMBER',
+            'updated-at': now,
+          },
+        },
+        {
+          type: 'certification-center-invitations',
+          id: certificationCenterInvitation2.id.toString(),
+          attributes: {
+            email: 'sarah.pelle@example.net',
+            role: 'ADMIN',
+            'updated-at': now,
+          },
+        },
+      ]);
+    });
+  });
+
   describe('POST /api/admin/certification-centers/{certificationCenterId}/invitations', function () {
     let clock;
     const now = new Date('2021-05-01');
